requirements.txt
# Core dependencies
python-docx==0.8.11
Markdown>=3.4.0
beautifulsoup4==4.10.0
PyYAML>=5.4.1
requests>=2.25.1

# Test dependencies
pytest>=7.0.0
pytest-cov>=2.12.0  # 用于测试覆盖率报告
pytest-html>=3.1.1  # 用于生成HTML测试报告

# Automatically generated by https://github.com/damnever/pigar.

# D:\Projects\Python\markdocx\src\provider\parser\extension\ext_md_syntax.py: 1,2
# D:\Projects\Python\markdocx\src\test\test_md2html.py: 1
Markdown == 3.3.3

# D:\Projects\Python\markdocx\src\test\test_html2docx.py: 7
beautifulsoup4 == 4.10.0

# D:\Projects\Python\markdocx\src\test\test_docx.py: 1,2
# D:\Projects\Python\markdocx\src\test\test_html2docx.py: 12,13,14,15
python_docx == 0.8.11

